Split data files related API controller code in a base class so it can be re-used by other API controllers.
Also update affected code.
